Upgrade react-router to v6+
This was initially done to fix the percent sign issue when searching for
albums, but despite comments mentioning it should be fixed by now, it
continues to be broken in v6, so I had to resort to using the string
`PERCENT_SIGN` and replace accordingly. This is *not* a problem related
to encode/decodeURIComponent.

Fixes OPENSCROBBLER-RB
Fixes OPENSCROBBLER-AB